dot case is ready
