+++ src/main/java/com/example/file/vault/FileVaultApplication.java
@@ -1,0 +1,14 @@
+package com.example.file.vault;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FileVaultApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FileVaultApplication.class, args);
+	}
+
+
+}
