+++ src/main/java/com/example/file/vault/constants/FileVaultConstants.java
@@ -1,0 +1,8 @@
+package com.example.file.vault.constants;
+
+public class FileVaultConstants {
+    public static final String ZIP_NAME = "attachments.zip";
+    public static final int MAX_FILE_SIZE_MB = 15;
+    public static final double BYTES_TO_MB_MULTIPLIER = 0.00000095367432; // 1 / 1024 / 1024
+    public static final long MB_TO_BYTES_MULTIPLIER = 1048576; // 1024 * 1024
+}
