+++ src/main/java/com/example/file/vault/exception/BadFileTypeException.java
@@ -1,0 +1,7 @@
+package com.example.file.vault.exception;
+
+public class BadFileTypeException extends RuntimeException{
+    public BadFileTypeException(String message) {
+        super(message);
+    }
+}
