+++ src/main/java/com/example/file/vault/exception/CloseStreamException.java
@@ -1,0 +1,11 @@
+package com.example.file.vault.exception;
+
+public class CloseStreamException extends RuntimeException{
+    public CloseStreamException(String message) {
+        super(message);
+    }
+
+    public CloseStreamException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
