+++ src/main/java/com/example/file/vault/exception/EmptyFileListException.java
@@ -1,0 +1,7 @@
+package com.example.file.vault.exception;
+
+public class EmptyFileListException extends RuntimeException{
+    public EmptyFileListException(String message) {
+        super(message);
+    }
+}
