+++ src/main/java/com/example/file/vault/exception/EmptyFileNameException.java
@@ -1,0 +1,7 @@
+package com.example.file.vault.exception;
+
+public class EmptyFileNameException extends RuntimeException{
+    public EmptyFileNameException(String message) {
+        super(message);
+    }
+}
