+++ src/main/java/com/example/file/vault/exception/EmptyGetParamException.java
@@ -1,0 +1,11 @@
+package com.example.file.vault.exception;
+
+public class EmptyGetParamException extends RuntimeException {
+    public EmptyGetParamException(String message) {
+        super(message);
+    }
+
+    public EmptyGetParamException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
