+++ src/main/java/com/example/file/vault/exception/FileNotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.file.vault.exception;
+
+public class FileNotFoundException extends RuntimeException{
+    public FileNotFoundException(String message) {
+        super(message);
+    }
+}
