+++ src/main/java/com/example/file/vault/exception/IncorrectDateTypeFilterParamException.java
@@ -1,0 +1,11 @@
+package com.example.file.vault.exception;
+
+public class IncorrectDateTypeFilterParamException extends RuntimeException {
+    public IncorrectDateTypeFilterParamException(String message) {
+        super(message);
+    }
+
+    public IncorrectDateTypeFilterParamException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
