+++ src/main/java/com/example/file/vault/exception/IncorrectFilterParamException.java
@@ -1,0 +1,11 @@
+package com.example.file.vault.exception;
+
+public class IncorrectFilterParamException extends RuntimeException {
+    public IncorrectFilterParamException(String message) {
+        super(message);
+    }
+
+    public IncorrectFilterParamException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
