+++ src/main/java/com/example/file/vault/exception/PutNextEntryToZipException.java
@@ -1,0 +1,12 @@
+package com.example.file.vault.exception;
+
+public class PutNextEntryToZipException extends RuntimeException{
+
+    public PutNextEntryToZipException(String message) {
+        super(message);
+    }
+
+    public PutNextEntryToZipException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
