+++ src/main/java/com/example/file/vault/exception/ReadWriteStreamException.java
@@ -1,0 +1,11 @@
+package com.example.file.vault.exception;
+
+public class ReadWriteStreamException extends RuntimeException{
+    public ReadWriteStreamException(String message) {
+        super(message);
+    }
+
+    public ReadWriteStreamException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
