+++ src/main/java/com/example/file/vault/exception/TooLargeFileSizeException.java
@@ -1,0 +1,12 @@
+package com.example.file.vault.exception;
+
+public class TooLargeFileSizeException extends RuntimeException {
+
+    public TooLargeFileSizeException(String message) {
+        super(message);
+    }
+
+    public TooLargeFileSizeException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
