+++ src/main/java/com/example/file/vault/repository/DateBaseStorageRepository.java
@@ -1,0 +1,5 @@
+package com.example.file.vault.repository;
+
+public class DateBaseStorageRepository {
+    // TODO: in 3 part
+}
