+++ src/main/java/com/example/file/vault/service/DataBaseStorageService.java
@@ -1,0 +1,6 @@
+package com.example.file.vault.service;
+
+public class DataBaseStorageService {
+    // TODO: in 3 part
+
+}
