--- src/main/java/com/example/fileVault/FileVaultApplication.java
@@ -1,16 +1,0 @@
-package com.example.fileVault;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-@SpringBootApplication
-public class FileVaultApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FileVaultApplication.class, args);
-	}
-
-
-}
