--- src/main/java/com/example/fileVault/exception/BadFileTypeException.java
@@ -1,7 +1,0 @@
-package com.example.fileVault.exception;
-
-public class BadFileTypeException extends RuntimeException{
-    public BadFileTypeException(String message) {
-        super(message);
-    }
-}
