--- src/main/java/com/example/fileVault/exception/CantReadFileContentException.java
@@ -1,11 +1,0 @@
-package com.example.fileVault.exception;
-
-public class CantReadFileContentException extends RuntimeException {
-    public CantReadFileContentException(String message) {
-        super(message);
-    }
-
-    public CantReadFileContentException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
