--- src/main/java/com/example/fileVault/exception/CloseStreamException.java
@@ -1,11 +1,0 @@
-package com.example.fileVault.exception;
-
-public class CloseStreamException extends RuntimeException{
-    public CloseStreamException(String message) {
-        super(message);
-    }
-
-    public CloseStreamException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
