--- src/main/java/com/example/fileVault/exception/EmptyFileListException.java
@@ -1,7 +1,0 @@
-package com.example.fileVault.exception;
-
-public class EmptyFileListException extends RuntimeException{
-    public EmptyFileListException(String message) {
-        super(message);
-    }
-}
