--- src/main/java/com/example/fileVault/exception/EmptyFileNameException.java
@@ -1,7 +1,0 @@
-package com.example.fileVault.exception;
-
-public class EmptyFileNameException extends RuntimeException{
-    public EmptyFileNameException(String message) {
-        super(message);
-    }
-}
