--- src/main/java/com/example/fileVault/exception/EmptyGetParamException.java
@@ -1,11 +1,0 @@
-package com.example.fileVault.exception;
-
-public class EmptyGetParamException extends RuntimeException {
-    public EmptyGetParamException(String message) {
-        super(message);
-    }
-
-    public EmptyGetParamException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
