--- src/main/java/com/example/fileVault/exception/FileNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.fileVault.exception;
-
-public class FileNotFoundException extends RuntimeException{
-    public FileNotFoundException(String message) {
-        super(message);
-    }
-}
