--- src/main/java/com/example/fileVault/exception/IncorrectDateTypeFilterParamException.java
@@ -1,11 +1,0 @@
-package com.example.fileVault.exception;
-
-public class IncorrectDateTypeFilterParamException extends RuntimeException {
-    public IncorrectDateTypeFilterParamException(String message) {
-        super(message);
-    }
-
-    public IncorrectDateTypeFilterParamException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
