--- src/main/java/com/example/fileVault/exception/IncorrectFilterParamException.java
@@ -1,11 +1,0 @@
-package com.example.fileVault.exception;
-
-public class IncorrectFilterParamException extends RuntimeException {
-    public IncorrectFilterParamException(String message) {
-        super(message);
-    }
-
-    public IncorrectFilterParamException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
