--- src/main/java/com/example/fileVault/exception/PutNextEntryToZipException.java
@@ -1,12 +1,0 @@
-package com.example.fileVault.exception;
-
-public class PutNextEntryToZipException extends RuntimeException{
-
-    public PutNextEntryToZipException(String message) {
-        super(message);
-    }
-
-    public PutNextEntryToZipException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
