--- src/main/java/com/example/fileVault/exception/TooLargeFileSizeException.java
@@ -1,12 +1,0 @@
-package com.example.fileVault.exception;
-
-public class TooLargeFileSizeException extends RuntimeException {
-
-    public TooLargeFileSizeException(String message) {
-        super(message);
-    }
-
-    public TooLargeFileSizeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
