--- src/main/java/com/example/fileVault/repository/DateBaseStorageRepository.java
@@ -1,5 +1,0 @@
-package com.example.fileVault.repository;
-
-public class DateBaseStorageRepository {
-    // TODO: in 3 part
-}
