--- src/main/java/com/example/fileVault/service/DataBaseStorageService.java
@@ -1,10 +1,0 @@
-package com.example.fileVault.service;
-
-import com.example.fileVault.entity.FileEntity;
-
-import java.util.UUID;
-
-public class DataBaseStorageService {
-    // TODO: in 3 part
-
-}
