--- src/test/java/com/example/fileVault/FileVaultApplicationTests.java
@@ -1,23 +1,0 @@
-package com.example.fileVault;
-
-import com.example.fileVault.controller.FileController;
-import com.example.fileVault.repository.FileSystemFileRepository;
-import com.example.fileVault.service.FileService;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-@SpringBootTest
-class FileVaultApplicationTests {
-
-	@Autowired
-	private FileController controller;
-
-	@Test
-	void contextLoads() {
-		assertThat(controller).isNotNull();
-	}
-
-}
